refactor: rename client file and update to LiteLLM
